+++ DotNut/BlindedMessage.cs
@@ -4,7 +4,7 @@ namespace DotNut;
 
 public class BlindedMessage
 {
-    [JsonPropertyName("amount")] public int Amount { get; set; }
+    [JsonPropertyName("amount")] public ulong Amount { get; set; }
     [JsonPropertyName("id")] public KeysetId Id { get; set; }
     [JsonPropertyName("B_")] public PubKey B_ { get; set; }
     [JsonPropertyName("witness")][JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)] public string? Witness { get; set; }
